feat: split encoder_decoder_Fn to two files
